+++ src/main/java/thunder/hack/features/hud/impl/TotemCounter.java
@@ -34,12 +34,12 @@ public class TotemCounter extends HudElement {
         float factor = Math.abs(angle < 0 ? angle / 15f : 0f);
 
         context.getMatrices().pushMatrix();
-        context.getMatrices().translate(xPos, yPos);
+        context.getMatrices().translate((float) (xPos), (float) (yPos));
         context.getMatrices().rotate(-((float) Math.toRadians(-Render2DEngine.interpolateFloat(prevAngle, angle, Render3DEngine.getTickDelta()))));
-        context.getMatrices().translate(-xPos, -yPos);
+        context.getMatrices().translate((float) (-xPos), (float) (-yPos));
 
 
-        context.getMatrices().translate(xPos - 36, yPos - 9);
+        context.getMatrices().translate((float) (xPos - 36), (float) (yPos - 9));
         context.drawItem(Items.TOTEM_OF_UNDYING.getDefaultStack(), 0, 0);
         context.getMatrices().translate(-(xPos - 36), -(yPos - 9));
